refactor: split AnimatedFABExample into separete files
