+++ example/src/Examples/AnimatedFABExample/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AnimatedFABExample';
